+++ src/main/java/org/jolokia/docker/maven/StartMojo.java
@@ -129,7 +129,7 @@ public class StartMojo extends AbstractDockerMojo {
                 long waited = WaitUtil.wait(wait.getTime(), checkers.toArray(new WaitUtil.WaitChecker[0]));
                 log.info(imageConfig.getDescription() + ": Waited " + StringUtils.join(logOut.toArray(), " and ") + " " + waited + " ms");
             } catch (TimeoutException exp) {
-                String desc = imageConfig.getDescription() + ": Timeout after " + wait.getTime() + " ms while waiting on " +
+                String desc = imageConfig.getDescription() + ": Timeout after " + wait.getTime() + " ms while waiting " +
                               StringUtils.join(logOut.toArray(), " and ");
                 log.error(desc);
                 throw new MojoExecutionException(desc);