+++ src/main/java/org/jolokia/docker/maven/access/hc/unix/UnixSocketClientBuilder.java
@@ -1,23 +1,23 @@
 package org.jolokia.docker.maven.access.hc.unix;
 
-import java.net.*;
+import java.net.InetAddress;
+import java.net.UnknownHostException;
 
 import org.apache.http.config.Registry;
 import org.apache.http.config.RegistryBuilder;
 import org.apache.http.conn.DnsResolver;
-import org.apache.http.conn.HttpClientConnectionManager;
 import org.apache.http.conn.socket.ConnectionSocketFactory;
 import org.apache.http.impl.client.*;
 import org.apache.http.impl.conn.PoolingHttpClientConnectionManager;
 
 public class UnixSocketClientBuilder {
 
-
-    public CloseableHttpClient build(String unixSocketPath) {
+    public CloseableHttpClient build(String unixSocketPath, int maxConnections) {
         final HttpClientBuilder httpBuilder = HttpClients.custom();
         final Registry<ConnectionSocketFactory> registry = buildRegistry(unixSocketPath);
         final DnsResolver dnsResolver = nullDnsResolver();
-        final HttpClientConnectionManager manager = new PoolingHttpClientConnectionManager(registry, dnsResolver);
+        final PoolingHttpClientConnectionManager manager = new PoolingHttpClientConnectionManager(registry, dnsResolver);
+        manager.setDefaultMaxPerRoute(maxConnections);
         httpBuilder.setConnectionManager(manager);
         return httpBuilder.build();
     }